+++ lib/td_dd_web/controllers/grant_request_controller.ex
@@ -16,6 +16,14 @@ defmodule TdDdWeb.GrantRequestController do
     end
   end
 
+  def index(conn, %{"user" => "me"} = params) do
+    with claims <- conn.assigns[:current_resource],
+         {:ok, grant_requests} <-
+           Requests.list_grant_requests(claims, Map.put(params, "user_id", claims.user_id)) do
+      render(conn, "index.json", grant_requests: grant_requests)
+    end
+  end
+
   def index(conn, %{} = params) do
     with claims <- conn.assigns[:current_resource],
          {:can, true} <- {:can, can?(claims, list(GrantRequest))},
@@ -26,8 +34,8 @@ defmodule TdDdWeb.GrantRequestController do
 
   def show(conn, %{"id" => id}) do
     with claims <- conn.assigns[:current_resource],
-         {:can, true} <- {:can, can?(claims, show(GrantRequest))} do
-      grant_request = Requests.get_grant_request!(id)
+         grant_request <- Requests.get_grant_request!(id),
+         {:can, true} <- {:can, can?(claims, show(grant_request))} do
       render(conn, "show.json", grant_request: grant_request)
     end
   end